+++ R/class-npv.R
@@ -40,7 +40,7 @@ class(npv) <- c("class_metric", "function")
 #' @export
 npv.data.frame <- function(data, truth, estimate,
                            prevalence = NULL,
-                           estimator = NULL, na.rm = TRUE, ...) {
+                           estimator = NULL, na_rm = TRUE, ...) {
 
   metric_summarizer(
     metric_nm = "npv",
@@ -49,7 +49,7 @@ npv.data.frame <- function(data, truth, estimate,
     truth = !!enquo(truth),
     estimate = !!enquo(estimate),
     estimator = estimator,
-    na.rm = na.rm,
+    na_rm = na_rm,
     ... = ...,
     metric_fn_options = list(prevalence = prevalence)
   )
@@ -82,7 +82,7 @@ npv.matrix <- function(data, prevalence = NULL, estimator = NULL, ...) {
 #' @rdname npv
 npv_vec <- function(truth, estimate,
                     prevalence = NULL,
-                    estimator = NULL, na.rm = TRUE, ...) {
+                    estimator = NULL, na_rm = TRUE, ...) {
 
   estimator <- finalize_estimator(truth, estimator)
 
@@ -101,7 +101,7 @@ npv_vec <- function(truth, estimate,
     metric_impl = npv_impl,
     truth = truth,
     estimate = estimate,
-    na.rm = na.rm,
+    na_rm = na_rm,
     estimator = estimator,
     cls = "factor",
     ...,