+++ R/class-spec.R
@@ -39,7 +39,7 @@ class(spec) <- c("class_metric", "function")
 #' @export
 #' @rdname spec
 spec.data.frame <- function(data, truth, estimate,
-                            estimator = NULL, na.rm = TRUE, ...) {
+                            estimator = NULL, na_rm = TRUE, ...) {
 
   metric_summarizer(
     metric_nm = "spec",
@@ -48,7 +48,7 @@ spec.data.frame <- function(data, truth, estimate,
     truth = !!enquo(truth),
     estimate = !!enquo(estimate),
     estimator = estimator,
-    na.rm = na.rm,
+    na_rm = na_rm,
     ... = ...
   )
 
@@ -78,7 +78,7 @@ spec.matrix <- function(data, estimator = NULL, ...) {
 
 #' @export
 #' @rdname spec
-spec_vec <- function(truth, estimate, estimator = NULL, na.rm = TRUE,...) {
+spec_vec <- function(truth, estimate, estimator = NULL, na_rm = TRUE,...) {
 
   estimator <- finalize_estimator(truth, estimator)
 
@@ -97,7 +97,7 @@ spec_vec <- function(truth, estimate, estimator = NULL, na.rm = TRUE,...) {
     metric_impl = spec_impl,
     truth = truth,
     estimate = estimate,
-    na.rm = na.rm,
+    na_rm = na_rm,
     estimator = estimator,
     cls = "factor",
     ...