+++ R/prob-lift_curve.R
@@ -84,7 +84,7 @@ lift_curve <- function(data, ...) {
 
 #' @rdname lift_curve
 #' @export
-lift_curve.data.frame <- function(data, truth, ..., na.rm = TRUE) {
+lift_curve.data.frame <- function(data, truth, ..., na_rm = TRUE) {
 
   estimate <- dots_to_estimate(data, !!! enquos(...))
   truth <- enquo(truth)
@@ -99,7 +99,7 @@ lift_curve.data.frame <- function(data, truth, ..., na.rm = TRUE) {
     lift_curve_vec(
       truth = rlang::eval_tidy(truth, data = .),
       estimate = rlang::eval_tidy(estimate, data = .),
-      na.rm = na.rm
+      na_rm = na_rm
     )
   )
 
@@ -113,10 +113,10 @@ lift_curve.data.frame <- function(data, truth, ..., na.rm = TRUE) {
   res
 }
 
-lift_curve_vec <- function(truth, estimate, na.rm = TRUE, ...) {
+lift_curve_vec <- function(truth, estimate, na_rm = TRUE, ...) {
 
   # tibble result, possibly grouped
-  res <- gain_curve_vec(truth, estimate, na.rm)
+  res <- gain_curve_vec(truth, estimate, na_rm)
 
   res <- dplyr::mutate(res, .lift = .percent_found / .percent_tested)